4.21

+++ viikko4/blogilista/controllers/blogs.js
@@ -12,11 +12,33 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (!request.token || !decodedToken) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (decodedToken.id.toString() !== blog.user.toString()) {
+      response.status(401).send('You cannot delete blogs made by others')
+    } else {
+
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+
+    }
   } catch (exception) {
-    console.log(exception)
-    response.status(400).send({ error: 'malformatted id' })
+    if (exception.name === 'JsonWebTokenError') {
+      response.status(401).json({ error: exception.message })
+    } else {
+      console.log(exception)
+      response.status(500).json({ error: `something went wrong...` })
+    }
   }
 })
